init_dotfiles: skip already symlinked bin scripts

+++ init_dotfiles.sh
@@ -20,11 +20,15 @@ if [ "$REPLY" = "y" ] ; then
 
     for script in $(ls ~/dotfiles/bin); do
         if [ -f ~/bin/$script ] ; then
-            read -n 1 -p "~/bin/$script already exists. Delete file? <y/n>"
-            echo ""
-            if [ "$REPLY" = "y" ] ; then
-                rm -rf ~/bin/$script
-                ln -s ~/dotfiles/bin/$script ~/bin/$script
+            if [[ -z $(ls -l ~/bin/$script | grep dotfiles) ]] ; then
+                read -n 1 -p "~/bin/$script already exists. Delete file? <y/n>"
+                echo ""
+                if [ "$REPLY" = "y" ] ; then
+                    rm -rf ~/bin/$script
+                    ln -s ~/dotfiles/bin/$script ~/bin/$script
+                fi
+            else
+                echo "~/bin/$script already symlinked to dotfiles"
             fi
         else
             ln -s ~/dotfiles/bin/$script ~/bin/$script
